# backend/openapi.yaml
openapi: "3.0.3"
info:
  title: Galaxy of Beauty API
  description: |
    REST API for the Galaxy of Beauty (جالكسي بيوتي) platform — a secure marketplace
    connecting female customers with vetted female technicians for at-home
    beauty and grooming services in Saudi Arabia.

    **Base URL:** `/api`

    **Authentication:** JWT Bearer token in `Authorization` header.
  version: "1.0.0"
  contact:
    name: Galaxy of Beauty
    email: support@galaxyofbeauty.sa
servers:
  - url: http://localhost:4000/api
    description: Local Development
  - url: https://api.galaxyofbeauty.sa/api
    description: Production
...
security:
  - bearerAuth: []
paths:
...
  /services:
    get:
      tags: [Services]
      summary: List services (filtered, paginated, localized)
      security: []
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [newest, price, -price, popularity, duration]
  /services/{id}:
    get:
      tags: [Services]
      summary: Get service detail (with variants, add-ons, tags, technicians)
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer

  # ── Bookings ────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...